backend/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
simple-websocket>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
